+++ HISTORY.rst
@@ -4,8 +4,9 @@ History
 ===========
 1.0.18
 -----------
-
-
+ * add: put_image()
+ * change: capture_screen()'s 3rd and 4th parameter to width and height
+ * add: more FillStyle consts
 
 1.0.17
 -----------